deleted reconstruct-stringify-walker.cpp
